fixed adaboost
