+++ sml/tree/BUILD.bazel
@@ -20,8 +20,3 @@ py_library(
     name = "tree",
     srcs = ["tree.py"],
 )
-
-py_library(
-    name = "tree_w",
-    srcs = ["tree_w.py"],
-)
